requirements.txt
transformers
torch
praw
tqdm
python-dotenv
cohere